extend terrafrom aliases for powershell

+++ powershell/aliases.ps1
@@ -367,16 +367,19 @@ ${function:Set-Power-Balanced} = { powercfg.exe /SETACTIVE SCHEME_BALANCED }
 ${function:Set-Power-Min} = { powercfg.exe /SETACTIVE SCHEME_MAX }
 
 # Terraform
-${function:terrafrom} = { terraform.exe @args }
-${function:t} = { terraform.exe @args }
-${function:ti} = { terraform.exe init @args }
-${function:tw} = { terraform.exe workspace @args }
-${function:twd} = { terraform.exe workspace delete @args }
-${function:twn} = { terraform.exe workspace new @args }
-${function:twl} = { terraform.exe workspace list @args }
-${function:tws} = { terraform.exe workspace select @args }
-${function:tp} = { terraform.exe plan -out terraform.plan @args }
-${function:ta} = { terraform.exe apply terraform.plan @args }
+if (Get-Command kitchen.bat -ErrorAction SilentlyContinue | Test-Path) {
+    ${function:terrafrom}   = { terraform.exe @args }
+    ${function:t}           = { terraform.exe @args }
+    ${function:ta}          = { terraform.exe apply terraform.plan @args }
+    ${function:ti}          = { terraform.exe init @args }
+    ${function:tp}          = { if ($args[0] -And -Not $args[1]) {terraform.exe plan -out terraform.plan --var-file=$args.tfvars} else {terraform.exe plan -out terraform.plan @args}}
+    ${function:tpd}         = { if ($args[0] -And -Not $args[1]) {terraform.exe plan -destroy -out terraform.plan --var-file=$args.tfvars} else {terraform.exe plan -destroy -out terraform.plan @args}}
+    ${function:tw}          = { terraform.exe workspace @args }
+    ${function:twd}         = { terraform.exe workspace delete @args }
+    ${function:twn}         = { terraform.exe workspace new @args }
+    ${function:twl}         = { terraform.exe workspace list @args }
+    ${function:tws}         = { terraform.exe workspace select @args }
+}
 
 ################################################################################
 ### >> Align:
